% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{add_file}
\alias{add_file}
\title{Adds files to the users current home directory from the aqm extdata directory}
\usage{
add_file(fl = "sleep.csv")
}
\arguments{
\item{fl}{file to add. Run aqm_data() to see options}
}
\description{
Useful for demonstrating reading and saving data
}
\examples{
aqm_data()


}
