\name{aqm-package}
\alias{aqm-package}
\alias{aqm}
\docType{package}
\title{
\packageTitle{aqm}
}
\description{
\packageDescription{aqm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aqm}
\packageIndices{aqm}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{aqm}

Maintainer: \packageMaintainer{aqm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
