% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{beta_likert}
\alias{beta_likert}
\title{beta_likert}
\usage{
beta_likert(n = 100, mean = 0.5, sd = 0.5)
}
\arguments{
\item{n}{}

\item{mean}{}

\item{sd}{}
}
\description{
A function to simulate responses on the Likert scale
with probabilities which follow a beta distribution
}
\examples{

## Symetrical around neutral
table(beta_likert(mean=0.5,sd=0.4))

## Marmite response, but on the side of agree overall

table(beta_likert(mean=0.7,sd=0.8))

## Strongly disagree

table(beta_likert(mean=0.1,sd=0.2))

}
