% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{beta_likert_vec}
\alias{beta_likert_vec}
\title{Title}
\usage{
beta_likert_vec(x = seq(0.1, 0.9, length = 100), sd = 0.5)
}
\arguments{
\item{x}{}

\item{sd}{}
}
\description{
Takes a vector of means between 0.1 and 0.9 and returns 
random Likert values.
The larger the given sd the greater the random variability around the mean
So the function can be used to simulate correlated responses if
from a latent factor.
}
\examples{


}
