% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{dmbeta}
\alias{dmbeta}
\title{Find density of reparameterised beta distribution}
\usage{
dmbeta(x = seq(0, 1, length = 1000), mean = 0.5, sd = 0.2)
}
\arguments{
\item{x}{}

\item{mean}{}

\item{sd}{}
}
\description{
Find density of reparameterised beta distribution
}
