% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{dynamite}
\alias{dynamite}
\title{Quickly forms dynamite plot}
\usage{
dynamite(g0, ...)
}
\arguments{
\item{g0}{ggplot object with x and y aesthetics set}

\item{...}{}
}
\description{
Quickly forms dynamite plot
}
