% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{f_met}
\alias{f_met}
\title{Download historical monthly records from met office}
\usage{
f_met(nm = "hurn", skip = 7)
}
\arguments{
\item{nm}{Name of the station. Defaults to Hurn}

\item{skip}{Number of lines to skip. Defaults to 7}
}
\description{
Download historical monthly records from met office
}
