% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{lm_eq}
\alias{lm_eq}
\title{Linear model equation}
\usage{
lm_eq(mod = reg_mod, dps = 1)
}
\arguments{
\item{mod}{A linear regression with}

\item{dps}{Decimal places}
}
\value{
y = ax +bx
}
\description{
Prints the regression equation for a linear model
}
