% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{median_cl_boot}
\alias{median_cl_boot}
\title{Provide bootstrapped confidence intervals for the median}
\usage{
median_cl_boot(x, conf = 0.95)
}
\arguments{
\item{x}{}

\item{conf}{}
}
\description{
Useful in a range of situations as a replacement for non-parametric tests
}
\examples{

d <- data.frame(group = rep(c("A", "B", "C", "D"), each = 100), mean = rep(c(10, 12, 15, 20), each = 100))
d$value <- d$mean + rnorm(400, 0, 5)
g0 <- ggplot(d, aes(x = group, y = value))
g_box <- g0 + geom_boxplot(fill = "grey", colour = "black", notch = TRUE) + theme_bw()
g_box <- g_box + stat_summary(fun.data = median_cl_boot, geom = "errorbar", colour = "red") + stat_summary(fun.y = median, geom = "point", colour = "red")
g_box



}
