% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot function
From R cookbook 
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{}

\item{plotlist}{}

\item{file}{}

\item{cols}{}

\item{layout}{}
}
\description{
Multiplot function
From R cookbook 
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
