% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{rand_likert}
\alias{rand_likert}
\title{rand_likert: Form a vector of likert responses with given probabilities}
\usage{
rand_likert(n = 1000, p = c(1, 2, 3, 5, 10))
}
\arguments{
\item{n}{Number of responses required}

\item{p}{A vector of probabilities. Do not have to sum to 1}
}
\value{
A vector of Likert responses
}
\description{
rand_likert: Form a vector of likert responses with given probabilities
}
\examples{

q1<-rand_likert(n=100,p=c(1,1,1,1,1)) # Equal probabilites
table(q1)

}
