% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{rmbeta}
\alias{rmbeta}
\title{Reparameterise the beta distribution in terms of mean and sd}
\usage{
rmbeta(n = 1000, mean = 0.5, sd = 0.2)
}
\arguments{
\item{n}{}

\item{mean}{}

\item{sd}{}
}
\description{
Produces a vector of random numbers between 0 and 1 with a given mean and sd
}
\examples{

hist(rmbeta(n=1000,mean=0.1))

}
