% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqm.R
\name{Rsq}
\alias{Rsq}
\title{Prints the R Squared value as % from a fitted regression model}
\usage{
Rsq(mod = reg_mod, dps = 1)
}
\arguments{
\item{mod}{Model}

\item{dps}{Number of decimal places}
}
\description{
Prints the R Squared value as % from a fitted regression model
}
